All tables are finished ("no data" handling, correct titles)

+++ src/main/java/comp3111/covid/TableEntry.java
@@ -4,16 +4,20 @@ import java.time.LocalDate;
 
 public class TableEntry {
 	String name;
-	Integer data1;
-	Integer data2;
+
+	Float data1;
+	Float data2;
+
+
 	LocalDate data3;
+
 	
-	TableEntry(String name, Integer data1, Integer data2){
+	TableEntry(String name, Float data1, Float data2){
 		this.name=name;
 		this.data1=data1;
 		this.data2=data2;
 	}
-	TableEntry(String name, LocalDate data3, Integer data2){
+	TableEntry(String name, LocalDate data3, Float data2){
 		this.name = name;
 		this.data3 = data3;
 		this.data2 = data2;
@@ -24,11 +28,11 @@ public class TableEntry {
 		return name;
 	}
 	
-	public Integer getData1() {
+	public Float getData1() {
 		return data1;
 	}
 	
-	public Integer getData2() {
+	public Float getData2() {
 		return data2;
 	}
